Split content out into own class
