+++ osu.Game/Screens/Select/Carousel/SetPanelContent.cs
@@ -1,0 +1,93 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using System.Collections.Generic;
+using System.Linq;
+using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Localisation;
+using osu.Game.Beatmaps.Drawables;
+using osu.Game.Graphics;
+using osu.Game.Graphics.Sprites;
+using osuTK;
+
+namespace osu.Game.Screens.Select.Carousel
+{
+    public class SetPanelContent : CompositeDrawable
+    {
+        private readonly CarouselBeatmapSet carouselSet;
+
+        public SetPanelContent(CarouselBeatmapSet carouselSet)
+        {
+            this.carouselSet = carouselSet;
+
+            // required to ensure we load as soon as any part of the panel comes on screen
+            RelativeSizeAxes = Axes.Both;
+        }
+
+        [BackgroundDependencyLoader]
+        private void load()
+        {
+            var beatmapSet = carouselSet.BeatmapSet;
+
+            InternalChild = new FillFlowContainer
+            {
+                // required to ensure we load as soon as any part of the panel comes on screen
+                RelativeSizeAxes = Axes.Both,
+                Direction = FillDirection.Vertical,
+                Padding = new MarginPadding { Top = 5, Left = 18, Right = 10, Bottom = 10 },
+                Children = new Drawable[]
+                {
+                    new OsuSpriteText
+                    {
+                        Text = new LocalisedString((beatmapSet.Metadata.TitleUnicode, beatmapSet.Metadata.Title)),
+                        Font = OsuFont.GetFont(weight: FontWeight.Bold, size: 22, italics: true),
+                        Shadow = true,
+                    },
+                    new OsuSpriteText
+                    {
+                        Text = new LocalisedString((beatmapSet.Metadata.ArtistUnicode, beatmapSet.Metadata.Artist)),
+                        Font = OsuFont.GetFont(weight: FontWeight.SemiBold, size: 17, italics: true),
+                        Shadow = true,
+                    },
+                    new FillFlowContainer
+                    {
+                        Direction = FillDirection.Horizontal,
+                        AutoSizeAxes = Axes.Both,
+                        Margin = new MarginPadding { Top = 5 },
+                        Children = new Drawable[]
+                        {
+                            new BeatmapSetOnlineStatusPill
+                            {
+                                Origin = Anchor.CentreLeft,
+                                Anchor = Anchor.CentreLeft,
+                                Margin = new MarginPadding { Right = 5 },
+                                TextSize = 11,
+                                TextPadding = new MarginPadding { Horizontal = 8, Vertical = 2 },
+                                Status = beatmapSet.Status
+                            },
+                            new FillFlowContainer<DifficultyIcon>
+                            {
+                                AutoSizeAxes = Axes.Both,
+                                Spacing = new Vector2(3),
+                                ChildrenEnumerable = getDifficultyIcons(),
+                            },
+                        }
+                    }
+                }
+            };
+        }
+
+        private const int maximum_difficulty_icons = 18;
+
+        private IEnumerable<DifficultyIcon> getDifficultyIcons()
+        {
+            var beatmaps = carouselSet.Beatmaps.ToList();
+
+            return beatmaps.Count > maximum_difficulty_icons
+                ? (IEnumerable<DifficultyIcon>)beatmaps.GroupBy(b => b.Beatmap.Ruleset).Select(group => new FilterableGroupedDifficultyIcon(group.ToList(), group.Key))
+                : beatmaps.Select(b => new FilterableDifficultyIcon(b));
+        }
+    }
+}
